+++ tests/test_database_url.py
@@ -69,6 +69,11 @@ def test_database_url_options():
     u = DatabaseURL("postgresql://localhost/mydatabase?pool_size=20&ssl=true")
     assert u.options == {"pool_size": "20", "ssl": "true"}
 
+    u = DatabaseURL(
+        "mysql+asyncmy://username:password@/testsuite?unix_socket=/tmp/mysqld/mysqld.sock"
+    )
+    assert u.options == {"unix_socket": "/tmp/mysqld/mysqld.sock"}
+
 
 def test_replace_database_url_components():
     u = DatabaseURL("postgresql://localhost/mydatabase")