+++ CHANGELOG.md
@@ -4,6 +4,26 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## 0.7.0 (Dec 18th, 2022)
+
+### Fixed
+
+* Fixed breaking changes in SQLAlchemy cursor; supports `>=1.4.42,<1.5` (#513).
+* Wrapped types in `typing.Optional` where applicable (#510).
+
+## 0.6.2 (Nov 7th, 2022)
+
+### Changed
+
+* Pinned SQLAlchemy `<=1.4.41` to avoid breaking changes (#520).
+
+## 0.6.1 (Aug 9th, 2022)
+
+### Fixed
+
+* Improve typing for `Transaction` (#493)
+* Allow string indexing into Record (#501)
+
 ## 0.6.0 (May 29th, 2022)
 
 * Dropped Python 3.6 support (#458)